simplify the column index implementation
